extract logging configuration to utils and rename init class and controller
