+++ libs/neuron_utils.py
@@ -5,6 +5,15 @@ from geppettoJupyter.geppetto_comm import GeppettoCoreAPI as G
 
 from geppettoJupyter.geppetto_comm import GeppettoJupyterModelSync
 
+def configure_logging():
+    logger = logging.getLogger()
+    fhandler = logging.FileHandler(filename='neuron.log', mode='a')
+    formatter = logging.Formatter(
+        '%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    fhandler.setFormatter(formatter)
+    logger.addHandler(fhandler)
+    logger.setLevel(logging.DEBUG)
+
 def extractGeometries():
     GeppettoJupyterModelSync.current_model.addGeometries(extractMorphology())
